+++ eslint.config.mjs
@@ -83,7 +83,6 @@ export default [
 
       'scripts/updateVersion.ts',
       'src/actions/BackupModalActions.tsx',
-      'src/actions/CategoriesActions.ts',
 
       'src/actions/CreateWalletActions.tsx',
 
@@ -105,7 +104,7 @@ export default [
       'src/actions/SoundActions.ts',
       'src/actions/TokenTermsActions.tsx',
       'src/actions/TransactionExportActions.tsx',
-      'src/actions/WalletActions.tsx',
+
       'src/actions/WalletListActions.tsx',
       'src/actions/WalletListMenuActions.tsx',
       'src/app.ts',
@@ -175,8 +174,6 @@ export default [
 
       'src/components/modals/CategoryModal.tsx',
 
-      'src/components/modals/ContactListModal.tsx',
-      'src/components/modals/ContactsPermissionModal.tsx',
       'src/components/modals/CountryListModal.tsx',
       'src/components/modals/DateModal.tsx',
       'src/components/modals/FiatListModal.tsx',
@@ -185,7 +182,7 @@ export default [
 
       'src/components/modals/FundAccountModal.tsx',
       'src/components/modals/GradientFadeout.tsx',
-      'src/components/modals/HelpModal.tsx',
+
       'src/components/modals/InsufficientFeesModal.tsx',
       'src/components/modals/ListModal.tsx',
       'src/components/modals/LoanWelcomeModal.tsx',
@@ -199,7 +196,6 @@ export default [
       'src/components/modals/ScanModal.tsx',
       'src/components/modals/StateProvinceListModal.tsx',
 
-      'src/components/modals/SwapVerifyTermsModal.tsx',
       'src/components/modals/TextInputModal.tsx',
       'src/components/modals/TransferModal.tsx',
       'src/components/modals/WalletListMenuModal.tsx',
@@ -223,7 +219,7 @@ export default [
       'src/components/navigation/TransactionDetailsTitle.tsx',
       'src/components/notification/NotificationCenterCard.tsx',
       'src/components/progress-indicators/AccountSyncBar.tsx',
-      'src/components/progress-indicators/CancellableProcessingScene.tsx',
+
       'src/components/progress-indicators/FullScreenLoader.tsx',
       'src/components/progress-indicators/LoadingSplashScreen.tsx',
 
@@ -242,7 +238,6 @@ export default [
       'src/components/scenes/ConfirmScene.tsx',
       'src/components/scenes/CreateWalletAccountSelectScene.tsx',
       'src/components/scenes/CreateWalletAccountSetupScene.tsx',
-      'src/components/scenes/CreateWalletCompletionScene.tsx',
 
       'src/components/scenes/CreateWalletImportOptionsScene.tsx',
       'src/components/scenes/CreateWalletImportScene.tsx',
@@ -282,8 +277,6 @@ export default [
       'src/components/scenes/Loans/LoanStatusScene.tsx',
 
       'src/components/scenes/ManageTokensScene.tsx',
-      'src/components/scenes/MigrateWalletCalculateFeeScene.tsx',
-      'src/components/scenes/MigrateWalletCompletionScene.tsx',
 
       'src/components/scenes/NotificationCenterScene.tsx',
       'src/components/scenes/NotificationScene.tsx',
@@ -299,7 +292,6 @@ export default [
       'src/components/scenes/SwapSuccessScene.tsx',
       'src/components/scenes/SweepPrivateKeyCalculateFeeScene.tsx',
       'src/components/scenes/SweepPrivateKeyCompletionScene.tsx',
-      'src/components/scenes/SweepPrivateKeyProcessingScene.tsx',
 
       'src/components/scenes/TransactionDetailsScene.tsx',
 
@@ -411,7 +403,7 @@ export default [
       'src/controllers/loan-manager/redux/actions.ts',
       'src/controllers/loan-manager/util/waitForLoanAccountSync.ts',
       'src/hooks/animations/useFadeAnimation.ts',
-      'src/hooks/redux/useContactThumbnail.ts',
+
       'src/hooks/useAbortable.ts',
       'src/hooks/useAccountSyncRatio.tsx',
       'src/hooks/useAsyncEffect.ts',