+++ src/components/scenes/MigrateWalletCompletionScene.tsx
@@ -42,7 +42,7 @@ interface MigrateWalletTokenItem extends MigrateWalletItem {
   tokenId: string
 }
 
-const MigrateWalletCompletionComponent = (props: Props) => {
+const MigrateWalletCompletionComponent: React.FC<Props> = props => {
   const { navigation, route } = props
   const { migrateWalletList } = route.params
 
@@ -54,12 +54,11 @@ const MigrateWalletCompletionComponent = (props: Props) => {
 
   const sortedMigrateWalletListBundles = React.useMemo(() => {
     return migrateWalletList.reduce((bundles: MigrateWalletItem[][], asset) => {
-      const { createWalletIds } = asset
-      const walletId = createWalletIds[0]
+      const { createWalletId: walletId } = asset
 
       // Find the bundle with the main currency at the end of it
       const index = bundles.findIndex(
-        bundle => walletId === bundle[0].createWalletIds[0]
+        bundle => walletId === bundle[0].createWalletId
       )
 
       if (index === -1) {
@@ -98,7 +97,7 @@ const MigrateWalletCompletionComponent = (props: Props) => {
   const handleItemStatus = (
     item: MigrateWalletItem,
     status: 'complete' | 'error'
-  ) => {
+  ): void => {
     setItemStatus(currentState => ({ ...currentState, [item.key]: status }))
     const index = sortedMigrateWalletList.findIndex(
       asset => asset.key === item.key
@@ -119,8 +118,7 @@ const MigrateWalletCompletionComponent = (props: Props) => {
       const migrationPromises = []
       for (const bundle of sortedMigrateWalletListBundles) {
         const mainnetItem = bundle[bundle.length - 1]
-        const { createWalletIds } = mainnetItem
-        const oldWalletId = createWalletIds[0]
+        const { createWalletId: oldWalletId } = mainnetItem
 
         securityCheckedWallets[oldWalletId] ??= {
           checked: false,
@@ -135,7 +133,7 @@ const MigrateWalletCompletionComponent = (props: Props) => {
         const newWalletName = `${oldWalletName}${lstrings.migrate_wallet_new_fragment}`
 
         // Create new wallet
-        const createNewWalletPromise = async () => {
+        const createNewWalletPromise = async (): Promise<void> => {
           const previouslyCreatedWalletInfo = account.allKeys.find(
             keys =>
               keys.migratedFromWalletId === oldWalletId &&
@@ -342,9 +340,7 @@ const MigrateWalletCompletionComponent = (props: Props) => {
   const renderRow = useHandler(
     (data: ListRenderItemInfo<MigrateWalletItem>) => {
       const { item } = data
-      const { createWalletIds } = item
-
-      const walletId = createWalletIds[0]
+      const { createWalletId: walletId } = item
       const wallet = currencyWallets[walletId]
       if (wallet == null) return null
       return (